Sensor: Invensense: 6515: fix step counter corrupting event queue
Fixed a sensor event queue corruption bug caused in some cases
when a step event is read.  The bug would either cause the step
event to get overwritten before being delivered to Sensor Services,
resulting in a lost sensor event and an older, already sent sensor
event to be resent to sensor services.

Bug: 25618874
Change-Id: I9d4c87b42f88113275a8ac613ae87ae9ec20d6ea